// src/app/data/models/request/projects/projects.ts
export interface UpdateProjectDetail {
    key: string;
    name: string;
    description: string;
    enabled: boolean;
}

export interface ObjectNewProject {
    clientId: string,
    clientName: string,
    newProject: NewProject,
}

export interface NewProject { 
    key: string;
    name: string;
    description: string;
}

export interface ProjectListPaginator{
    page: number;
    size: number;
}